cleaning code for simple hash and removing race condition for cache

+++ consistent_web_main/consistent_hash/simple_hash.go
@@ -9,16 +9,16 @@ import (
 type SimpleHash struct {
 	orderedKeys    []string
 	nodeMap        map[string]ServerNode
-	sizeInclRepls  uint32
+	sizeInclRepls  int
 	mux            sync.RWMutex
 }
 
 func NewSimpleHash(nodeMap map[string]ServerNode) *SimpleHash {
-	var size uint32 = 0
+	var size int = 0
 	orderedKeys := make([]string, 0)
 
 	for ip, node := range nodeMap  {
-		size += (uint32) (node.Replicas)
+		size += node.Replicas
 		orderedKeys = append(orderedKeys, ip)
 	}
 	h := &SimpleHash{
@@ -34,9 +34,9 @@ func (h *SimpleHash) InsertNode(ip_address string, replica_count int) {
 	h.mux.Lock()
 	defer h.mux.Unlock()
 	if entry, ok := h.nodeMap[ip_address]; ok {
-		h.sizeInclRepls -= (uint32) (entry.Replicas)
+		h.sizeInclRepls -= entry.Replicas
 		entry.Replicas = replica_count
-		h.sizeInclRepls += (uint32) (entry.Replicas)
+		h.sizeInclRepls += entry.Replicas
 
 		// Note that this changes the distribution for other keys after the ip address
 		// this is fine because regardless the other keys will be changed
@@ -45,14 +45,14 @@ func (h *SimpleHash) InsertNode(ip_address string, replica_count int) {
 		timestamp := time.Now().Add(60 * time.Second)
 		h.nodeMap[ip_address] = ServerNode{IP: ip_address, Timestamp: timestamp, Replicas: replica_count}
 		h.orderedKeys = append(h.orderedKeys, ip_address)
-		h.sizeInclRepls += (uint32) (replica_count)
+		h.sizeInclRepls += replica_count
 	}
 }
 
 func (h *SimpleHash) DeleteNode(ip string) {
 	h.mux.Lock()
 	defer h.mux.Unlock()
-	h.sizeInclRepls -= (uint32) (h.nodeMap[ip].Replicas)
+	h.sizeInclRepls -= h.nodeMap[ip].Replicas
 	index := sort.Search(len(h.orderedKeys), func(i int) bool {
 		return h.orderedKeys[i] >= ip
 	})
@@ -66,10 +66,10 @@ func (h *SimpleHash) DeleteNode(ip string) {
 func (h *SimpleHash) ValueLookup(value string) string {
 	h.mux.RLock()
 	defer h.mux.RUnlock()
-	replica_idx := getTrieKey(value) % h.sizeInclRepls
-	var tempSize uint32 = 0
+	replica_idx := (int) (getTrieKey(value)) % h.sizeInclRepls
+	var tempSize int = 0
 	for _, ip := range h.orderedKeys {
-		tempSize += (uint32) (h.nodeMap[ip].Replicas)
+		tempSize += h.nodeMap[ip].Replicas
 		if replica_idx < tempSize {
 			return ip
 		}
